# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API YocoShort",
        "description": "Documentación oficial de YocoShort",
        "contact": {
            "email": "sisabuestandavidesteban@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.yocoshort.com",
            "description": "Servidor de Producción en Render"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local para desarrollo"
        }
    ],
    "paths": {
...
        "/api/links": {
            "get": {
                "tags": [
                    "User Links"
                ],
                "summary": "Listar todos mis enlaces",
                "operationId": "524d62542e36f8f16e2546a2c966962c",
                "responses": {
                    "200": {
                        "description": "Lista de enlaces obtenida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "short_code": {
                                                        "type": "string",
                                                        "example": "promo-2025"
                                                    },
                                                    "original_url": {
                                                        "type": "string",
                                                        "example": "https://google.com"
                                                    },
                                                    "clicks": {
                                                        "type": "integer",
                                                        "example": 120
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Links"
                ],
                "summary": "Crear un nuevo enlace corto",
                "operationId": "1662829072423f1f1feffec526b5941e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "original_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://www.google.com"
                                    },
                                    "etiquetas": {
                                        "type": "string",
                                        "example": "Campaña Verano"
                                    },
                                    "custom_alias": {
                                        "type": "string",
                                        "example": "promo-2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enlace creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Link creado exitosamente."
                                        },
                                        "short_url": {
                                            "type": "string",
                                            "example": "http://marca.yocoshort.com/promo-2025"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación o falta configurar dominio"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/links/{id}": {
            "put": {
                "tags": [
                    "User Links"
                ],
                "summary": "Actualizar alias (short code)",
                "operationId": "8f1df907e6a5ca85677664934e6dbb32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del enlace",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "short_code": {
                                        "type": "string",
                                        "example": "nuevo-alias"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alias actualizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Link no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Links"
                ],
                "summary": "Eliminar un enlace",
                "operationId": "4db677de606cf59fe13b945effe2e2d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del enlace",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Link eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}